requirements.txt
emoji==2.2.0
jieba==0.42.1
matplotlib==3.6.2
numpy==1.19.5
pandas==1.4.4
Pillow==9.4.0
scikit_learn==1.2.0
tqdm==4.64.1
wordcloud==1.8.2.2
